Move stores to own directory
